Move PantsDaemonStats into run_tracker.py
and remove unused methods

[ci skip-rust]

[ci skip-build-wheels]
